// src/app/Model/bardetails.ts
export interface BarDetails {
  CustomerCode: string;
  ProjectCode: string;
  JobID: number;
  BBSID: number;
  BarID: number;
  BarSort: number;
  Cancelled: boolean;
  BarCAB: boolean;
  BarSTD: boolean;
  ElementMark: string;
  BarMark: string;
  BarType: string;
  BarSize: number;
  BarMemberQty: number;
  BarEachQty: number;
  BarTotalQty: number;
  BarShapeCode: string;
  A: string;
  B: string;
  C: string;
  D: string;
  E: string;
  F: string;
  G: string;
  H: string;
  I: string;
  J: string;
  K: string;
  L: string;
  M: string;
  N: string;
  O: string;
  P: string;
  Q: string;
  R: string;
  S: string;
  T: string;
  U: string;
  V: string;
  W: string;
  X: string;
  Y: string;
  Z: string;
  BarLength: string;
  BarWeight: number;
  Remarks: string;
  shapeParameters: string;
  shapeLengthFormula: string;
  shapeParaValidator: string;
  shapeTransportValidator: string;
  shapeParType: string;
  shapeDefaultValue: string;
  shapeHeightCheck: string;
  shapeAutoCalcFormula1: string;
  shapeAutoCalcFormula2: string;
  shapeAutoCalcFormula3: string;
  shapeTransport: any;
  PinSize: number;
  TeklaGUID: string;
  PartGUID: string;
  UpdateDate: Date;
  UpdateBy:string
}